+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 202409.04
 
 - Fixed wastewater sample writing bug.
+- Added regex exclusion for KitTypeReagentRole to trim down Bacteria Positive Control list.
 
 ## 202409.03